update sdk version

+++ internal/cmd/kafka/command_mirror.go
@@ -31,7 +31,7 @@ type listMirrorWrite struct {
 	MirrorStatus             string
 	StatusTimeMs             int32
 	NumPartition             int32
-	MaxPerPartitionMirrorLag int64
+	MaxPerPartitionMirrorLag int32
 }
 
 type describeMirrorWrite struct {
@@ -41,7 +41,7 @@ type describeMirrorWrite struct {
 	MirrorStatus          string
 	StatusTimeMs          int32
 	Partition             int32
-	PartitionMirrorLag    int64
+	PartitionMirrorLag    int32
 	LastSourceFetchOffset int64
 }
